+++ src/VirtoCommerce.Build/PlatformTools/Modules/Azure/AzureBlobModuleItem.cs
@@ -1,0 +1,13 @@
+using PlatformTools.Modules;
+
+namespace PlatformTools.Modules.Azure
+{
+    internal class AzureBlobModuleItem : ModuleItem
+    {
+        public AzureBlobModuleItem(string id, string version) : base(id, version)
+        {
+
+        }
+        public string BlobName { get; set; }
+    }
+}