+++ src/VirtoCommerce.Build/PlatformTools/Modules/Azure/AzureCliToolSettings.cs
@@ -1,0 +1,11 @@
+using System;
+using Nuke.Common.Tooling;
+
+namespace PlatformTools.Modules.Azure
+{
+    [Serializable]
+    internal class AzureCliToolSettings : ToolSettings
+    {
+        public override Action<OutputType, string> ProcessLogger => ProcessTasks.DefaultLogger;
+    }
+}