+++ src/VirtoCommerce.Build/PlatformTools/Modules/Gitlab/GitlabJobArtifactsModuleItem.cs
@@ -1,0 +1,13 @@
+using PlatformTools.Modules;
+
+namespace PlatformTools.Modules.Gitlab;
+
+internal class GitlabJobArtifactsModuleItem : ModuleItem
+{
+    public GitlabJobArtifactsModuleItem(string id, string version) : base(id, version)
+    {
+    }
+
+    public string JobId { get; set; }
+    public string ArtifactName { get; set; }
+}