+++ src/VirtoCommerce.Build/PlatformTools/Modules/ModuleItem.cs
@@ -1,0 +1,14 @@
+namespace PlatformTools.Modules
+{
+    public class ModuleItem
+    {
+        public ModuleItem(string id, string version)
+        {
+            Id = id;
+            Version = version;
+        }
+
+        public string Id { get; set; }
+        public string Version { get; set; }
+    }
+}